payment status button

+++ backend/OrderKraft/src/main/java/com/dto/OrderDTO.java
@@ -1,5 +1,6 @@
 package com.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
@@ -10,7 +11,7 @@ public class OrderDTO {
 @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy")
 private LocalDate  order_date;//needs to be formatted, if necessary!
 private String status;
-private long total_amount;
+private BigDecimal total_amount;
 private long procurement_officer_id;
 
 public String getStatus() {
@@ -27,10 +28,10 @@ public void setOrder_date(LocalDate  order_date) {
 	this.order_date = order_date;
 }
 
-public long getTotal_amount() {
+public BigDecimal getTotal_amount() {
 	return total_amount;
 }
-public void setTotal_amount(long total_amount) {
+public void setTotal_amount(BigDecimal total_amount) {
 	this.total_amount = total_amount;
 }
 public long getProcurement_officer_id() {
@@ -44,7 +45,7 @@ public void setProcurement_officer_id(long procurement_officer_id) {
 public OrderDTO() {}
 
 //all-args constr.
-public OrderDTO(LocalDate order_date, String status, long total_amount, long procurement_officer_id) {
+public OrderDTO(LocalDate order_date, String status, BigDecimal total_amount, long procurement_officer_id) {
 	super();
 	this.order_date = order_date;
 	this.status = status;
